rename cli_entry.py to upload.py
