+++ iridauploader/core/__init__.py
@@ -1,4 +1,4 @@
 from iridauploader.core import logger
-from iridauploader.core import cli_entry
+from iridauploader.core import upload
 from iridauploader.core import exit_return
-from iridauploader.core.cli_entry import VERSION_NUMBER
+from iridauploader.core.upload import VERSION_NUMBER
