+++ iridauploader/core/upload.py
@@ -16,6 +16,7 @@ from iridauploader.model import DirectoryStatus
 from . import api_handler, parsing_handler, logger, exit_return, upload_helpers
 
 VERSION_NUMBER = "0.5.0"
+# TODO rename this file to upload.py
 
 
 def upload_run_single_entry(directory, force_upload=False, upload_mode=None):
