+++ scripts/update_version.sh
@@ -2,7 +2,7 @@
 # This script updates the version number in the various locations they need to be changed including:
 #   setup.py : for pypi / pip
 #   windows-installer.cfg: for pynsist windows builds
-#   cli_entry.py : for version identifiers in the main code base
+#   upload.py : for version identifiers in the main code base
 # Use example:
 # $./scripts/update_version.sh 0.4.2
 # Use the first argument as the new version number
@@ -14,6 +14,6 @@ sed -i "s/version=.*\,/version='$newversion',/" "$setuppy"
 wininstall=windows-installer.cfg
 sed -i -z "s/version=.\..\../version=$newversion/" "$wininstall"
 
-clientry=iridauploader/core/cli_entry.py
+clientry=iridauploader/core/upload.py
 sed -i "s/VERSION_NUMBER = \".\..\..\"/VERSION_NUMBER = \"$newversion\"/" "$clientry"
